【step09】修复材质文件编译后生成冗余数据的问题

+++ source/step09/Tools/ScriptCompiler/Include/T3DScriptCompiler.h
@@ -215,7 +215,7 @@ namespace Tiny3D
             char        magic[4];       /// 文件标识
             uint32_t    subtype;
             uint32_t    version;        /// 文件版本号
-            size_t      fileSize;       /// 文件大小
+            uint32_t    fileSize;       /// 文件大小
         };
 
     public:
